Add TypeScript support to src files
This commit consists of a series of changes that convert the src
JavaScript (JS) files to TypeScript (TS). The conversion process
involved renaming all files, updating import and export statements,
and finally backporting various types.

Babel was previously used to transcompile files directly on the src
folder. It is now instead redirected to operate on the output of the TSC
compiler.

PR: https://github.com/identity-com/credential-commons/pull/210
ISSUE: IDCOM-2324

Reviewed-by: Julian Spring <springjulian@gmail.com>
Reviewed-by: William Brooks <william@identity.org>
Signed-off-by: Tighe Barris <4658684+tbarri@users.noreply.github.com>